update tests; small refactors
